Fix/be post sh

See merge request s07-webmobile2-sub2/moggozi!361

+++ backend/restapiserver/src/main/java/com/JJP/restapiserver/controller/stage/PostController.java
@@ -24,7 +24,7 @@ public class PostController {
 
     // post 글 쓰기
     @PostMapping
-    private int save(@RequestBody PostSaveRequestDto postSaveRequestDto, HttpServletRequest request){
+    private Long save(@RequestBody PostSaveRequestDto postSaveRequestDto, HttpServletRequest request){
         Long member_id = jwtUtils.getUserIdFromJwtToken(request.getHeader("Authorization"));
 
         return postService.savePost(postSaveRequestDto, member_id);
